requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
APScheduler==3.10.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
Pillow==10.2.0
Werkzeug==2.3.7
SQLAlchemy==2.0.25
alembic==1.13.1
python-dateutil==2.8.2
pytz==2024.1
gunicorn==21.2.0 